Get rid of API-level services in authorization microservice
